Split and review model
