+++ test/Common/Evolved/Model.cs
@@ -1,0 +1,62 @@
+﻿/*
+*  Copyright (c) 2022-2026 MASES s.r.l.
+*
+*  Licensed under the Apache License, Version 2.0 (the "License");
+*  you may not use this file except in compliance with the License.
+*  You may obtain a copy of the License at
+*
+*  http://www.apache.org/licenses/LICENSE-2.0
+*
+*  Unless required by applicable law or agreed to in writing, software
+*  distributed under the License is distributed on an "AS IS" BASIS,
+*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+*  See the License for the specific language governing permissions and
+*  limitations under the License.
+*
+*  Refer to LICENSE for more information.
+*/
+
+
+using MASES.EntityFrameworkCore.KNet.Serialization;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Diagnostics;
+using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace MASES.EntityFrameworkCore.KNet.Test.Common.Model.Evolved
+{
+    [PrimaryKey("BlogId")]
+    [Table("Blog", Schema = "Simple")]
+    public class Blog
+    {
+        public int BlogId { get; set; }
+        public int Rating { get; set; }
+        public DateTime? Date { get; set; }
+
+        public List<Post> Posts { get; set; }
+
+        public override string ToString()
+        {
+            return $"BlogId: {BlogId} Rating: {Rating} Date: {Date}";
+        }
+    }
+
+    [PrimaryKey("PostId")]
+    [Table("Post", Schema = "Simple")]
+    public class Post
+    {
+        public int PostId { get; set; }
+        public string Title { get; set; }
+        public string Content { get; set; }
+
+        public int BlogId { get; set; }
+        public Blog Blog { get; set; }
+
+        public override string ToString()
+        {
+            return $"PostId: {PostId} Title: {Title} Content: {Content} BlogId: {BlogId}";
+        }
+    }
+}
