+++ test/Common/Model.cs
@@ -16,595 +16,27 @@
 *  Refer to LICENSE for more information.
 */
 
-using Javax.Smartcardio;
-using MASES.EntityFrameworkCore.KNet.Serialization;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Diagnostics;
-using Org.W3c.Dom.Ls;
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
 namespace MASES.EntityFrameworkCore.KNet.Test.Common.Model.Base
 {
-    [PrimaryKey("BlogId")]
-    [Table("Blog", Schema = "Simple")]
-    public class Blog
-    {
-        public int BlogId { get; set; }
-        public string Url { get; set; }
-        public int Rating { get; set; }
-
-        public List<Post> Posts { get; set; }
-
-        public override string ToString()
-        {
-            return $"BlogId: {BlogId} Url: {Url} Rating: {Rating}";
-        }
-    }
-
-    [PrimaryKey("PostId")]
-    [Table("Post", Schema = "Simple")]
-    public class Post
-    {
-        public int PostId { get; set; }
-        public string Title { get; set; }
-        public string Content { get; set; }
-
-        public int BlogId { get; set; }
-        public Blog Blog { get; set; }
-
-        public override string ToString()
-        {
-            return $"PostId: {PostId} Title: {Title} Content: {Content} BlogId: {BlogId}";
-        }
-    }
+    // moved into specific file
 }
 
 namespace MASES.EntityFrameworkCore.KNet.Test.Common.Model.Evolved
 {
-    [PrimaryKey("BlogId")]
-    [Table("Blog", Schema = "Simple")]
-    public class Blog
-    {
-        public int BlogId { get; set; }
-        public int Rating { get; set; }
-        public DateTime? Date { get; set; }
-
-        public List<Post> Posts { get; set; }
-
-        public override string ToString()
-        {
-            return $"BlogId: {BlogId} Rating: {Rating} Date: {Date}";
-        }
-    }
-
-    [PrimaryKey("PostId")]
-    [Table("Post", Schema = "Simple")]
-    public class Post
-    {
-        public int PostId { get; set; }
-        public string Title { get; set; }
-        public string Content { get; set; }
-
-        public int BlogId { get; set; }
-        public Blog Blog { get; set; }
-
-        public override string ToString()
-        {
-            return $"PostId: {PostId} Title: {Title} Content: {Content} BlogId: {BlogId}";
-        }
-    }
+    // moved into specific file
 }
 
 namespace MASES.EntityFrameworkCore.KNet.Test.Common.Model.ReducedComplex
 {
-    [PrimaryKey("BlogId")]
-    [Table("BlogComplex", Schema = "ReducedComplexTest")]
-    public class BlogComplex
-    {
-        public int BlogId { get; set; }
-        public string Url { get; set; }
-        public int Rating { get; set; }
-
-        public TaxInfo TaxInfo { get; set; }
-
-        [Required]
-        public TaxInfoExtended TaxInfoExtended { get; set; }
-
-        public override string ToString()
-        {
-            return $"BlogId: {BlogId} Url: {Url} Rating: {Rating}";
-        }
-    }
-
-    [ComplexType]
-    public class TaxInfoExtended
-    {
-        public TaxInfoExtended()
-        {
-        }
-
-        public TaxInfoExtended(string str)
-        {
-            try
-            {
-                var values = str.Split("_");
-                CodeExtended = int.Parse(values[0]);
-                PercentageExtended = decimal.Parse(values[1]);
-                NestedTaxInfoExtended = new NestedTaxInfoExtended(values[2]);
-            }
-            catch (Exception ex)
-            {
-                System.Diagnostics.Debug.WriteLine($"TaxInfoExtendedConverter.ConvertBack failed for input '{str}': {ex}");
-            }
-        }
-
-        public int CodeExtended { get; set; }
-        public decimal PercentageExtended { get; set; }
-        [Required]
-        public NestedTaxInfoExtended NestedTaxInfoExtended { get; set; }
-        public override string ToString()
-        {
-            return $"{CodeExtended}_{PercentageExtended}_{NestedTaxInfoExtended}";
-        }
-    }
-
-    [ComplexType]
-    public class NestedTaxInfoExtended
-    {
-        public NestedTaxInfoExtended()
-        {
-        }
-
-        public NestedTaxInfoExtended(string str)
-        {
-            try
-            {
-                var values = str.Split("$");
-                CodeExtended = int.Parse(values[0]);
-                PercentageExtended = decimal.Parse(values[1]);
-            }
-            catch (Exception ex)
-            {
-                System.Diagnostics.Debug.WriteLine($"TaxInfoExtendedConverter.ConvertBack failed for input '{str}': {ex}");
-            }
-        }
-
-        public int CodeExtended { get; set; }
-        public decimal PercentageExtended { get; set; }
-        public override string ToString()
-        {
-            return $"{CodeExtended}${PercentageExtended}";
-        }
-    }
-
-    [Owned]
-    [Table("TaxInfo", Schema = "ReducedComplexTest")]
-    public class TaxInfo
-    {
-        public int TaxInfoId { get; set; }
-        public char Code { get; set; }
-        public decimal Percentage { get; set; }
-    }
-
-    public class TaxInfoExtendedConverter : IComplexTypeConverter
-    {
-        public IEnumerable<Type> SupportedClrTypes => [typeof(TaxInfoExtended)];
-
-        public IDiagnosticsLogger<DbLoggerCategory.Infrastructure> Logging { get; set; }
-
-        public bool Convert(PreferredConversionType conversionType, ref object input)
-        {
-            if (input is TaxInfoExtended taxInfoExtended)
-            {
-                input = taxInfoExtended.ToString();
-                return true;
-            }
-            return false;
-        }
-
-        public bool ConvertBack(PreferredConversionType conversionType, ref object input)
-        {
-            if (input is string str)
-            {
-                input = new TaxInfoExtended(str);
-                return true;
-            }
-            return false;
-        }
-
-        public void Register(IDiagnosticsLogger<DbLoggerCategory.Infrastructure> logging)
-        {
-            Logging = logging;
-        }
-    }
+    // moved into specific file
 }
 
 namespace MASES.EntityFrameworkCore.KNet.Test.Common.Model.Complex
 {
-    [PrimaryKey("BlogId")]
-    [Table("BlogComplex", Schema = "ComplexTest")]
-    public class BlogComplex
-    {
-        public int BlogId { get; set; }
-        public string Url { get; set; }
-        public int Rating { get; set; }
-
-        public bool BooleanValue { get; set; }
-        public bool? NullableBooleanValue { get; set; }
-
-        // Nested complex type
-        public Pricing PricingInfo { get; set; }
-
-        public List<PostComplex> ComplexPosts { get; set; }
-
-        public override string ToString()
-        {
-            return $"BlogId: {BlogId} Url: {Url} Rating: {Rating} BooleanValue: {BooleanValue} NullableBooleanValue: {NullableBooleanValue}";
-        }
-    }
-
-    [PrimaryKey("PricingId")]
-    [Table("Pricing", Schema = "ComplexTest")]
-    public class Pricing
-    {
-        public int PricingId { get; set; }
-        public decimal BasePrice { get; set; }
-        public List<Discount> Discounts { get; set; } // Nested collection!
-        public TaxInfo Tax { get; set; } // Nested object!
-    }
-
-    [PrimaryKey("DiscountId")]
-    [Table("Discount", Schema = "ComplexTest")]
-    public class Discount
-    {
-        public int DiscountId { get; set; }
-        public string Code { get; set; }
-        public decimal Percentage { get; set; }
-        public DateRange Validity { get; set; } // Another nested object!
-    }
-
-    [PrimaryKey("DateRangeId")]
-    [Table("DateRange", Schema = "ComplexTest")]
-    public class DateRange
-    {
-        public int DateRangeId { get; set; }
-        public uint CurrentDiff { get; set; }
-        public DateTime Min { get; set; }
-        public DateTime Max { get; set; }
-    }
-
-    [Owned]
-    [Table("TaxInfo", Schema = "ComplexTest")]
-    public class TaxInfo
-    {
-        public int TaxInfoId { get; set; }
-        public char Code { get; set; }
-        public decimal Percentage { get; set; }
-        [Required]
-        public TaxInfoExtended TaxInfoExtended { get; set; }
-        [Required]
-        public TaxInfoExtended TaxInfoExtended2 { get; set; }
-        public int ExtraValue { get; set; } // used to check index consistency since the method GetIndex of IComplexProperty is zero-based 
-    }
-
-    [ComplexType]
-    public class TaxInfoExtended : IEquatable<TaxInfoExtended>
-    {
-        public TaxInfoExtended()
-        {
-        }
-
-        public TaxInfoExtended(string str)
-        {
-            try
-            {
-                var values = str.Split("_");
-                CodeExtended = int.Parse(values[0]);
-                PercentageExtended = decimal.Parse(values[1]);
-            }
-            catch (Exception ex)
-            {
-                System.Diagnostics.Debug.WriteLine($"TaxInfoExtendedConverter.ConvertBack failed for input '{str}': {ex}");
-            }
-        }
-
-        public int CodeExtended { get; set; }
-        public decimal PercentageExtended { get; set; }
-
-        public override string ToString()
-        {
-            return $"{CodeExtended}_{PercentageExtended}";
-        }
-        public override bool Equals(object obj)
-        {
-            if (obj is TaxInfoExtended no)
-            {
-                return Equals(no);
-            }
-            return base.Equals(obj);
-        }
-        public override int GetHashCode()
-        {
-            return HashCode.Combine(CodeExtended, PercentageExtended);
-        }
-
-        public bool Equals(TaxInfoExtended other)
-        {
-            return other.CodeExtended == CodeExtended
-                   && other.PercentageExtended == PercentageExtended;
-        }
-    }
-
-    public class TaxInfoExtendedConverter : IComplexTypeConverter
-    {
-        public IEnumerable<Type> SupportedClrTypes => [typeof(TaxInfoExtended)];
-
-        public IDiagnosticsLogger<DbLoggerCategory.Infrastructure> Logging { get; set; }
-
-        public bool Convert(PreferredConversionType conversionType, ref object input)
-        {
-            if (input is TaxInfoExtended taxInfoExtended)
-            {
-                input = taxInfoExtended.ToString();
-                return true;
-            }
-            return false;
-        }
-
-        public bool ConvertBack(PreferredConversionType conversionType, ref object input)
-        {
-            if (input is string str)
-            {
-                input = new TaxInfoExtended(str);
-                return true;
-            }
-            return false;
-        }
-
-        public void Register(IDiagnosticsLogger<DbLoggerCategory.Infrastructure> logging)
-        {
-            Logging = logging;
-        }
-    }
-
-    [PrimaryKey("PostId")]
-    [Table("PostComplex", Schema = "ComplexTest")]
-    public class PostComplex
-    {
-        public int PostId { get; set; }
-        public string Title { get; set; }
-        public string Content { get; set; }
-        public Guid Identifier { get; set; }
-        public DateTimeOffset CreationTime { get; set; }
-
-        public int BlogId { get; set; }
-        public BlogComplex Blog { get; set; }
-
-        public override string ToString()
-        {
-            return $"PostId: {PostId} Title: {Title} Content: {Content} BlogId: {Blog?.BlogId} CreationTime: {CreationTime} Identifier: {Identifier}";
-        }
-    }
+    // moved into specific file
 }
 
 namespace MASES.EntityFrameworkCore.KNet.Test.Common.Model.MultiLevelComplex
 {
-    [PrimaryKey("BlogId")]
-    [Table("BlogComplex", Schema = "MultiLevelComplexTest")]
-    public class BlogComplex
-    {
-        public int BlogId { get; set; }
-        public string Url { get; set; }
-        public int Rating { get; set; }
-
-        public bool BooleanValue { get; set; }
-        public bool? NullableBooleanValue { get; set; }
-
-        // Nested complex type
-        public Pricing PricingInfo { get; set; }
-
-        public List<PostComplex> ComplexPosts { get; set; }
-
-        public override string ToString()
-        {
-            return $"BlogId: {BlogId} Url: {Url} Rating: {Rating} BooleanValue: {BooleanValue} NullableBooleanValue: {NullableBooleanValue}";
-        }
-    }
-
-    [PrimaryKey("PricingId")]
-    [Table("Pricing", Schema = "MultiLevelComplexTest")]
-    public class Pricing
-    {
-        public int PricingId { get; set; }
-        public decimal BasePrice { get; set; }
-        public List<Discount> Discounts { get; set; } // Nested collection!
-        public TaxInfo Tax { get; set; } // Nested object!
-    }
-
-    [PrimaryKey("DiscountId")]
-    [Table("Discount", Schema = "MultiLevelComplexTest")]
-    public class Discount
-    {
-        public int DiscountId { get; set; }
-        public string Code { get; set; }
-        public decimal Percentage { get; set; }
-        public DateRange Validity { get; set; } // Another nested object!
-    }
-
-    [PrimaryKey("DateRangeId")]
-    [Table("DateRange", Schema = "MultiLevelComplexTest")]
-    public class DateRange
-    {
-        public int DateRangeId { get; set; }
-        public uint CurrentDiff { get; set; }
-        public DateTime Min { get; set; }
-        public DateTime Max { get; set; }
-    }
-
-    [Owned]
-    [Table("TaxInfo", Schema = "MultiLevelComplexTest")]
-    public class TaxInfo
-    {
-        public int TaxInfoId { get; set; }
-        public char Code { get; set; }
-        public decimal Percentage { get; set; }
-        [Required]
-        public TaxInfoExtended TaxInfoExtended { get; set; }
-        [Required]
-        public TaxInfoExtended TaxInfoExtended2 { get; set; }
-        public int ExtraValue { get; set; } // used to check index consistency since the method GetIndex of IComplexProperty is zero-based 
-    }
-
-    [ComplexType]
-    public class TaxInfoExtended : IEquatable<TaxInfoExtended>
-    {
-        public TaxInfoExtended()
-        {
-        }
-
-        public TaxInfoExtended(string str)
-        {
-            try
-            {
-                var values = str.Split("_");
-                CodeExtended = int.Parse(values[0]);
-                PercentageExtended = decimal.Parse(values[1]);
-                NestedTaxInfoExtended = new NestedTaxInfoExtended(values[2]);
-            }
-            catch (Exception ex)
-            {
-                System.Diagnostics.Debug.WriteLine($"TaxInfoExtendedConverter.ConvertBack failed for input '{str}': {ex}");
-            }
-        }
-
-        public int CodeExtended { get; set; }
-        public decimal PercentageExtended { get; set; }
-        [Required]
-        public NestedTaxInfoExtended NestedTaxInfoExtended { get; set; }
-        public override string ToString()
-        {
-            return $"{CodeExtended}_{PercentageExtended}_{NestedTaxInfoExtended}";
-        }
-        public override bool Equals(object obj)
-        {
-            if (obj is TaxInfoExtended no)
-            {
-                return Equals(no);
-            }
-            return base.Equals(obj);
-        }
-        public override int GetHashCode()
-        {
-            return HashCode.Combine(CodeExtended, PercentageExtended, NestedTaxInfoExtended);
-        }
-
-        public bool Equals(TaxInfoExtended other)
-        {
-            return other.CodeExtended == CodeExtended
-                   && other.PercentageExtended == PercentageExtended
-                   && other.NestedTaxInfoExtended.Equals(NestedTaxInfoExtended);
-        }
-    }
-
-    [ComplexType]
-    public class NestedTaxInfoExtended : IEquatable<NestedTaxInfoExtended>
-    {
-        public NestedTaxInfoExtended()
-        {
-        }
-
-        public NestedTaxInfoExtended(string str)
-        {
-            try
-            {
-                var values = str.Split("$");
-                CodeExtended = int.Parse(values[0]);
-                PercentageExtended = decimal.Parse(values[1]);
-            }
-            catch (Exception ex)
-            {
-                System.Diagnostics.Debug.WriteLine($"TaxInfoExtendedConverter.ConvertBack failed for input '{str}': {ex}");
-            }
-        }
-
-        public int CodeExtended { get; set; }
-        public decimal PercentageExtended { get; set; }
-        public override string ToString()
-        {
-            return $"{CodeExtended}${PercentageExtended}";
-        }
-        public override bool Equals(object obj)
-        {
-            if (obj is NestedTaxInfoExtended no)
-            {
-                return Equals(no);
-            }
-            return base.Equals(obj);
-        }
-        public override int GetHashCode()
-        {
-            return HashCode.Combine(CodeExtended, PercentageExtended);
-        }
-
-        public bool Equals(NestedTaxInfoExtended other)
-        {
-            return other.CodeExtended == CodeExtended
-                   && other.PercentageExtended == PercentageExtended;
-        }
-    }
-
-    public class TaxInfoExtendedConverter : IComplexTypeConverter
-    {
-        public IEnumerable<Type> SupportedClrTypes => [typeof(TaxInfoExtended)];
-
-        public IDiagnosticsLogger<DbLoggerCategory.Infrastructure> Logging { get; set; }
-
-        public bool Convert(PreferredConversionType conversionType, ref object input)
-        {
-            if (input is TaxInfoExtended taxInfoExtended)
-            {
-                input = taxInfoExtended.ToString();
-                return true;
-            }
-            return false;
-        }
-
-        public bool ConvertBack(PreferredConversionType conversionType, ref object input)
-        {
-            if (input is string str)
-            {
-                input = new TaxInfoExtended(str);
-                return true;
-            }
-            return false;
-        }
-
-        public void Register(IDiagnosticsLogger<DbLoggerCategory.Infrastructure> logging)
-        {
-            Logging = logging;
-        }
-    }
-
-    [PrimaryKey("PostId")]
-    [Table("PostComplex", Schema = "MultiLevelComplexTest")]
-    public class PostComplex
-    {
-        public int PostId { get; set; }
-        public string Title { get; set; }
-        public string Content { get; set; }
-        public Guid Identifier { get; set; }
-        public DateTimeOffset CreationTime { get; set; }
-
-        public int BlogId { get; set; }
-        public BlogComplex Blog { get; set; }
-
-        public override string ToString()
-        {
-            return $"PostId: {PostId} Title: {Title} Content: {Content} BlogId: {Blog?.BlogId} CreationTime: {CreationTime} Identifier: {Identifier}";
-        }
-    }
+    // moved into specific file
 }
